+++ lib/groups/fix.go
@@ -39,7 +39,7 @@ func Fix(simulate bool) error {
 		logs.Printcr("no fixes needed")
 	}
 	elapsed := time.Since(start).Seconds()
-	logs.Print(fmt.Sprintf(", time taken %.1f seconds\n", float64(elapsed)))
+	logs.Print(fmt.Sprintf(", time taken %.1f seconds\n", elapsed))
 	return nil
 }
